==> 6c_plot_Gene-Enh_pairs.ipynb <==
.libPaths("/home/mahat/.conda/envs/r422/lib/R/library")
.libPaths()

suppressMessages({
    library(tidyverse)
    library(ggExtra)
    library(matrixStats)
    library(plyranges)
    library(viridis)
    library(data.table)
    library(ggbio)
    library(ggridges)
    library(ggbeeswarm)
    library(ggpointdensity)
    library(Matrix)
});
source("./scGRO_functions.r");

options(
    repr.plot.width=6,
    repr.plot.height=4,
    jupyter.plot_mimetypes = "image/svg+xml"
);
theme_set(theme_classic() +
    theme(
        axis.title.x = element_text(color="black", size=14, face="bold"), 
        axis.title.y = element_text(color="black", size=14, face="bold"),
        axis.text = element_text(color="black", size=12, face="bold"),
        plot.title = element_text(face="bold", size=14, hjust = 0.5),
        axis.line = element_blank(),
        # axis.ticks = element_blank()
        panel.border = element_rect(colour = "grey", fill=NA, linewidth=1)
    )
);

# Get equation and r^2 as string
# https://groups.google.com/forum/#!topic/ggplot2/1TgH-kG5XMA

lm_eqn = function(x, y) {
    m = lm(y ~ x);
    eq = substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2, 
         list(a = format(unname(coef(m)[1]), digits = 2),
              b = format(unname(coef(m)[2]), digits = 2),
             r2 = format(summary(m)$r.squared, digits = 3)))
    as.character(as.expression(eq));
}

#features = read_bed("../data/groHMM_mES_BRsComb_LP-50_UTS10_features_customized_v2.bed");
# features = read_bed("../data/mES_BRsComb_dREGfiltered_features_customized_OSNenhancersPlusSEs_v1.bed");
features = read_bed("../data/dREG_refinedFeatures_mES_mm10_OSNcustomEnhancers_SEs.bed");
names(features) = features$name;
features$score=NULL;
features

# # select genes longer than 5kp from each feature
# # truncate genes longer than 30kb to 30kb
# features = features %>%
#     filter( width(features) >= 3000 | substr(name, 0, 2) != "GN" ) %>%
#     anchor_center() %>%
#     mutate(width=width-1000);
# length(features)

#resize 30+kb genes to 30kb
longf = which( width(features) > 30000 & substr(names(features), 0, 2) == "GN" );
features[longf] = features[longf] %>%
     resize( width = 30000, fix="start" );
features

scGRO  = readRDS("../data/scGROv2p8_consolidated.rds");
counts = readRDS( "../data/scGROv2p8_mapq3qc_filtered_counts.rds" );
# scGRO  = readRDS("../jay_m/data_jay_m/scGROv2p9_filtered_consolidated.rds");
# counts = readRDS( "../jay_m/data_jay_m/scGROv2p9_mapq3qc_filtered_counts.rds" );
# hist(colSums(counts))
# head(scGRO)
# head(counts)

# Merge experiment ID and cell barcode to create unique cell ID across experiments
scGRO = scGRO %>%
    filter( umiQC & plateQC & cellQC & countQC & miRQC ) %>%
    mutate(cellID = factor( paste( Exp, Plate, Cell, sep="-") ) ) %>%
    # filter( !(Exp == "Exp236" & Plate %in% paste0("c0", 5:8)) ) %>%
    # filter( !(Exp == "Exp260b") ) %>%
    resize(width=1, fix="end") %>%
    filter(cellID %in% colnames(counts)) %>%
    # subsetByOverlaps(features) %>%
    select( cellID );
names(scGRO) = NULL;
scGRO

scrambled = scGRO;
scrambled$cellID = droplevels(scrambled$cellID);
scrambled$cellID = sample(scrambled$cellID);

# load dREG peak calls and convert to GRanges
dREG = read.table("../data/PROseq_mES_BRsComb.dREG.peak.full.bed", header=F, stringsAsFactors=F);
colnames(dREG) = c("chr", "start", "end", "score", "pval", "center");
dREG = GRanges(dREG)
dREG

get_position_matrix = function(reads, window, binsize=100) {
    tss = promoters(window, upstream=0, downstream=1);
    nbins = ceiling(width(window)/binsize);
    reads = subsetByOverlaps(reads, window) %>%
        mutate( pos = distanceToNearest(., tss) ) %>%
        mutate( bin = factor(ceiling(pos.distance/binsize), levels=1:nbins) );
    cells = levels(reads$cellID);
    #out = Matrix( nrow=nbins, ncol=length(cells), sparse=T );
    out = stats::xtabs( ~ bin + cellID, data=reads, sparse=T );
    colnames(out) = cells;
    
    return(out);
}

get_position_table = function(reads, gene, enh) {
    gtss = promoters(gene, upstream=0, downstream=1);
    etss = promoters(enh,  upstream=0, downstream=1);
    
    greads = subsetByOverlaps(reads, gene) %>%
        mutate( pos = distanceToNearest(., gtss) ) %>%
        as.data.frame %>%
        select(cellID, pos=pos.distance);
    
    ereads = subsetByOverlaps(reads, enh ) %>%
        mutate( pos = distanceToNearest(., etss) ) %>%
        as.data.frame %>%
        select(cellID, pos=pos.distance);
    
    out = left_join(greads, ereads,
                    by=c("cellID"), suffix = c(".gene", ".enh"),
                    multiple="all") %>%
        na.omit %>%
        mutate(diff = pos.enh - pos.gene);
    #out = c(greads, ereads);
    #out = out %>%
    #    group_by(cellID) %>%
    #    mutate(pos.distance = pos.distance-pos.distance[1]) %>%
    #    ungroup;
    
    return(out);
}

# plot_position_plormn_table = function(reads, gene, enh) {
#     table = get_position_table(reads, gene, enh);
#     out = table %>%
#     ggplot(aes(x=diff/1000)) +
#     xlab("Enh - Gene Polymerase position") +
#     ylab(paste0("Number of cells", " - (", nrow(table), ")")) +
#     geom_histogram(binwidth=1) +
#     # xlim(-30, 30) +
#     ggtitle(paste(names(gene), "-", names(enh)));
    
#     return(out);
# }

plot_position_table = function(reads, gene, enh) {
    strand(enh) = "+"
    pl = get_position_table(reads, gene, enh);
    strand(enh) = "-"
    mn = get_position_table(reads, gene, enh);
    table = rbind(pl, mn);
    out = table %>%
    ggplot(aes(x=diff/1000)) +
    xlab("Enh - Gene Polymerase position") +
    ylab(paste0("Number of cells", " - (", nrow(table), ")")) +
    geom_histogram(binwidth=1) +
    # xlim(-30, 30) +
    ggtitle(paste(names(gene), "-", names(enh)));
    
    return(out);
}

### plot distance between Pol II in gene and enhancer in same cells and thier positions

geneA = features["GN-Hccs"]
geneB = features["GN-LOC101055907"]
# strand(geneA) = "+";
# strand(geneB) = "-"
plot_position_pairs( scGRO, geneA, geneB, dREG );
plot_position_table( scGRO, geneA, geneB );

gene = features["GN-Sox2"]
enh = features["Sox2_2kbDn"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Sox2"]
enh = features["Sox2_105kbDn"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Sox2"]
enh = features["Sox2_1MbDn"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Pou5f1"]
enh = features["Pou5f1_3kbUp"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Pou5f1"]
enh = features["Pou5f1_25kbUp"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Pou5f1"]
enh = features["Pou5f1_20kbUp"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

SEcorr = chr_corr %>%
    filter( Gene %in% query_SE & Enhancer %in% query_SE ) %>%
    filter( emp_p_pos < 0.05 & distance <= 1000000 ) %>%
    filter(pGene >= 0.05 & pEnh >= 0.05 )#; %>%

Gene	Enhancer	pGene	pEnh	expCotrans	obsCotrans	chisq	emp_p_pos	emp_p_neg	geneLength	enhsLength	geneStrand	enhsStrand	distance
GN-Sulf1	INT_STITCHED_45	0.07974301	0.12925170	0.010306919	0.022675737	0.014843198	0.000	1.000	10000	45151	+	*	341631
GN-Nr5a2	INT_STITCHED_466	0.05177627	0.05215420	0.002700349	0.019652305	0.106419118	0.000	1.000	10000	9328	-	*	4524
GN-Etl4	INT_STITCHED_746	0.06727135	0.14436886	0.009711888	0.019652305	0.010174323	0.000	1.000	10000	25279	+	*	87622
GN-Dnmt3b	INT_STITCHED_1210	0.07029478	0.05177627	0.003639601	0.008314437	0.006004527	0.020	0.980	10000	13465	+	*	766718
GN-Cbfa2t2	INT_STITCHED_1210	0.08578987	0.05177627	0.004441879	0.010959940	0.009564670	0.002	0.998	10000	13465	+	*	20294
GN-Gar1	INT_STITCHED_1732	0.05555556	0.18178382	0.010099101	0.018896447	0.007663385	0.004	0.996	8787	14351	-	*	287297
GN-Rad23b	INT_STITCHED_1973	0.07482993	0.10733182	0.008031633	0.015495087	0.006935470	0.016	0.984	10000	21823	+	*	106345
GN-Rcc2	INT_STITCHED_2292	0.15873016	0.08541194	0.013557451	0.023053666	0.006651552	0.032	0.968	10000	16431	+	*	856646
GN-Kazn	INT_STITCHED_2292	0.05442177	0.08541194	0.004648269	0.012849584	0.014470241	0.000	1.000	10000	16431	-	*	681281
GN-Med13l	INT_STITCHED_2745	0.05668934	0.05631141	0.003192257	0.007558579	0.005972191	0.018	0.982	10000	17725	+	*	128066
GN-Pml	INT_STITCHED_4555	0.09183673	0.05177627	0.004754963	0.010204082	0.006244610	0.015	0.985	10000	12081	-	*	21904
GN-Gclc	INT_STITCHED_4657	0.05895692	0.06764928	0.003988393	0.008314437	0.004692280	0.041	0.959	10000	14043	+	*	604801
GN-Ooep	INT_STITCHED_4657	0.05593348	0.06764928	0.003783860	0.023053666	0.098134025	0.000	1.000	10000	14043	-	*	19251
GN-Rnu3a	INT_STITCHED_4954	0.62055933	0.72637944	0.450761543	0.619425548	0.063109968	0.000	1.000	4975	853	+	*	480
GN-Herc4	INT_STITCHED_5044	0.07785336	0.07747543	0.006031723	0.012093726	0.006092435	0.037	0.963	10000	13918	+	*	352834
GN-Hnrnph3	INT_STITCHED_5044	0.06802721	0.07747543	0.005270438	0.013227513	0.012013243	0.001	0.999	10000	13918	-	*	132970
GN-Epn2	INT_STITCHED_5533	0.05177627	0.05668934	0.002935162	0.006802721	0.005096144	0.028	0.972	10000	6155	-	*	929207
GN-LOC108167924	INT_STITCHED_5711	0.06160242	0.07105064	0.004376891	0.047619048	0.427217376	0.000	1.000	10000	6487	-	*	4039
GN-Ddx24	INT_STITCHED_6151	0.05668934	0.05064248	0.002870889	0.006424792	0.004399414	0.042	0.958	10000	12131	-	*	720122
GN-Ptch1	INT_STITCHED_6460	0.06198035	0.06160242	0.003818139	0.009070295	0.007224759	0.011	0.989	10000	26208	-	*	390221
GN-Ctsl	INT_STITCHED_6460	0.06991686	0.06160242	0.004307047	0.009070295	0.005267768	0.031	0.969	10000	26208	-	*	414708
GN-Basp1	INT_STITCHED_7104	0.06916100	0.16439909	0.011370005	0.021541950	0.009100124	0.006	0.994	10000	13894	-	*	336396
GN-Myo10	INT_STITCHED_7104_2	0.08163265	0.08767952	0.007157512	0.035903250	0.115447595	0.000	1.000	10000	10035	+	*	6417
GN-Usp7	INT_STITCHED_7359	0.07936508	0.11602419	0.009208269	0.016628874	0.005979992	0.041	0.959	10000	11985	-	*	29694
GN-H2-M5	INT_STITCHED_7792	0.05139834	0.07369615	0.003787859	0.027966742	0.154340046	0.000	1.000	8167	13848	-	*	3732
GN-Wdr43	INT_STITCHED_7884	0.09939531	0.05064248	0.005033625	0.010204082	0.005311007	0.025	0.975	10000	10887	+	*	870702
GN-Myl12a	INT_STITCHED_7884	0.05064248	0.05064248	0.002564661	0.006424792	0.005809975	0.014	0.986	10000	10887	-	*	257021

gene = features["GN-Sulf1"]
enh = features["INT_STITCHED_45"]
# plot_2strand_position_pairs(scGRO, gene, enh, dREG);
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Sulf1"]
enh = features["INT_STITCHED_45"]
start(gene) = 12692262;
gene = resize(gene, width=30000, fix="start");
start(enh) = 13069651;
strand(enh) = "+";
enh = resize(enh, width=30000, fix="start");
plot_position_pairs( scGRO, gene, enh, dREG );
end(enh) = 13095010;
strand(enh) = "-";
enh = resize(enh, width=30000, fix="start");
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Slco5a1"]
enh = features["INT_STITCHED_45"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Nr5a2"]
enh = features["INT_STITCHED_466"]
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Etl4"]
enh = features["INT_STITCHED_746"];
strand(enh) = "+";
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Dnmt3b"]
enh = features["INT_STITCHED_1210"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Cbfa2t2"]
enh = features["INT_STITCHED_1210"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );	

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gar1"]
enh = features["INT_STITCHED_1732"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );	

plot_position_table(scGRO, gene, enh);

gene = features["GN-Rad23b"]
enh = features["INT_STITCHED_1973"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Rcc2"]
enh = features["INT_STITCHED_2292"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Kazn"]
enh = features["INT_STITCHED_2292"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Med13l"]
enh = features["INT_STITCHED_2745"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"	
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Pml"]
enh = features["INT_STITCHED_4555"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gclc"]
enh = features["INT_STITCHED_4657"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ooep"]
enh = features["INT_STITCHED_4657"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gtf3c6"]
enh = features["INT_STITCHED_4954"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Herc4"]
enh = features["INT_STITCHED_5044"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Hnrnph3"]
enh = features["INT_STITCHED_5044"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Epn2"]
enh = features["INT_STITCHED_5533"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-LOC108167924"]
enh = features["INT_STITCHED_5711"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ddx24"]
enh = features["INT_STITCHED_6151"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ptch1"]
enh = features["INT_STITCHED_6460"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ctsl"]
enh = features["INT_STITCHED_6460"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Basp1"]
enh = features["INT_STITCHED_7104"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Myo10"]
enh = features["INT_STITCHED_7104_2"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Usp7"]
enh = features["INT_STITCHED_7359"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-H2-M5"]
enh = features["INT_STITCHED_7792"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Wdr43"]
enh = features["INT_STITCHED_7884"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Myl12a"]
enh = features["INT_STITCHED_7884"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

SEcorr = chr_corr %>%
    filter( Gene %in% query_SE & Enhancer %in% query_SE ) %>%
    filter( emp_p_pos < 0.05 & distance <= 100000 ) %>%
    filter(pGene >= 0.025 & pEnh >= 0.025 )#; %>%

Gene	Enhancer	pGene	pEnh	expCotrans	obsCotrans	chisq	emp_p_pos	emp_p_neg	geneLength	enhsLength	geneStrand	enhsStrand	distance
GN-Nr5a2	INT_STITCHED_466	0.05177627	0.05215420	0.002700349	0.019652305	0.106419118	0.000	1.000	10000	9328	-	*	4524
GN-Aldh9a1	INT_STITCHED_559	0.04761905	0.11111111	0.005291005	0.012093726	0.008746356	0.002	0.998	10000	7615	+	*	82953
GN-Sde2	INT_STITCHED_611	0.04081633	0.03061224	0.001249479	0.003779289	0.005122085	0.021	0.979	10000	10397	+	*	73212
GN-Etl4	INT_STITCHED_746	0.06727135	0.14436886	0.009711888	0.019652305	0.010174323	0.000	1.000	10000	25279	+	*	87622
GN-Gpr107	INT_STITCHED_803	0.02910053	0.05555556	0.001616696	0.004913076	0.006721191	0.004	0.996	10000	12043	+	*	94577
GN-Csnk2a1	INT_STITCHED_1196	0.05177627	0.04913076	0.002543807	0.006424792	0.005921062	0.008	0.992	10000	6803	+	*	51180
GN-Cbfa2t2	INT_STITCHED_1210	0.08578987	0.05177627	0.004441879	0.010959940	0.009564670	0.002	0.998	10000	13465	+	*	20294
GN-Gm39962	INT_STITCHED_1210	0.03099017	0.05177627	0.001604555	0.022675737	0.276708841	0.000	1.000	10000	13465	-	*	5161
GN-Gm14261	INT_STITCHED_1300	0.02532124	0.05706727	0.001445014	0.015495087	0.136610815	0.000	1.000	10000	24445	+	*	2941
GN-Trim33	INT_STITCHED_1658	0.04648526	0.02569917	0.001194633	0.003779289	0.005592055	0.006	0.994	10000	10755	+	*	74910
GN-Slc37a3	INT_STITCHED_3045	0.03741497	0.03212396	0.001201917	0.003401361	0.004024864	0.043	0.957	10000	8611	-	*	63946
GN-Nanog	INT_STITCHED_3348	0.05933485	0.03325775	0.001973343	0.010582011	0.037555124	0.000	1.000	10000	17754	+	*	17463
GN-Etnk1	INT_STITCHED_3437	0.06386999	0.03325775	0.002124172	0.005291005	0.004721290	0.032	0.968	10000	15957	+	*	65582
GN-Pml	INT_STITCHED_4555	0.09183673	0.05177627	0.004754963	0.010204082	0.006244610	0.015	0.985	10000	12081	-	*	21904
GN-Ooep	INT_STITCHED_4657	0.05593348	0.06764928	0.003783860	0.023053666	0.098134025	0.000	1.000	10000	14043	-	*	19251
GN-Rnu3a	INT_STITCHED_4954	0.62055933	0.72637944	0.450761543	0.619425548	0.063109968	0.000	1.000	4975	853	+	*	480
GN-2410006H16Rik	INT_STITCHED_5533	0.03930461	0.05668934	0.002228153	0.005291005	0.004210245	0.037	0.963	10000	6155	+	*	93980
GN-E130012A19Rik	INT_STITCHED_5711	0.04006047	0.07105064	0.002846322	0.008314437	0.010504883	0.002	0.998	7417	6487	-	*	26642
GN-LOC108167924	INT_STITCHED_5711	0.06160242	0.07105064	0.004376891	0.047619048	0.427217376	0.000	1.000	10000	6487	-	*	4039
GN-Rara	INT_STITCHED_5719	0.02721088	0.04383976	0.001192919	0.004535147	0.009364003	0.002	0.998	10000	6794	+	*	5948
GN-Mapt	INT_STITCHED_5752	0.03061224	0.05442177	0.001665973	0.011715797	0.060624638	0.000	1.000	10000	14603	+	*	11145
GN-Ppp2r5c	INT_STITCHED_6188_2	0.03779289	0.05857899	0.002213870	0.005291005	0.004277020	0.046	0.954	10000	7795	+	*	6714
GN-LOC102638940	INT_STITCHED_6188_2	0.04421769	0.05857899	0.002590227	0.036659108	0.448103001	0.000	1.000	10000	7795	-	*	6632
GN-Ppp2r5c	INT_STITCHED_6188	0.03779289	0.05442177	0.002056756	0.005668934	0.006343888	0.010	0.990	10000	10256	+	*	11287
GN-Gm30363	INT_STITCHED_6709	0.03174603	0.13983371	0.004439165	0.023809524	0.084522821	0.000	1.000	10000	9670	-	*	4391
GN-Dmtn	INT_STITCHED_6887	0.03212396	0.04157218	0.001335463	0.009826153	0.053982623	0.000	1.000	10000	9364	-	*	2697
GN-Tsc22d1	INT_STITCHED_6904	0.03476946	0.05177627	0.001800233	0.034013605	0.576426148	0.000	1.000	10000	10591	+	*	5024
GN-Myo10	INT_STITCHED_7104_2	0.08163265	0.08767952	0.007157512	0.035903250	0.115447595	0.000	1.000	10000	10035	+	*	6417
GN-Usp7	INT_STITCHED_7359	0.07936508	0.11602419	0.009208269	0.016628874	0.005979992	0.041	0.959	10000	11985	-	*	29694
GN-H2-M5	INT_STITCHED_7792	0.05139834	0.07369615	0.003787859	0.027966742	0.154340046	0.000	1.000	8167	13848	-	*	3732
GN-Rab12	INT_STITCHED_7876_2	0.03590325	0.09448224	0.003392219	0.007558579	0.005117167	0.039	0.961	10000	10118	-	*	42871
GN-Gm36856	INT_STITCHED_7887	0.03061224	0.05479970	0.001677542	0.007558579	0.020617429	0.000	1.000	10000	11863	+	*	9416
GN-Gm20544	INT_STITCHED_8261	0.02721088	0.04497354	0.001223770	0.026832955	0.535909867	0.000	1.000

gene = features["GN-Nr5a2"]
enh = features["INT_STITCHED_466"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Aldh9a1"]
enh = features["INT_STITCHED_559"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Sde2"]
enh = features["INT_STITCHED_611"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Etl4"]
enh = features["INT_STITCHED_746"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gpr107"]
enh = features["INT_STITCHED_803"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Csnk2a1"]
enh = features["INT_STITCHED_1196"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Cbfa2t2"]
enh = features["INT_STITCHED_1210"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gm39962"]
enh = features["INT_STITCHED_1210"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gm14261"]
enh = features["INT_STITCHED_1300"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Trim33"]
enh = features["INT_STITCHED_1658"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Slc37a3"]
enh = features["INT_STITCHED_3045"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Nanog"]
enh = features["INT_STITCHED_3348"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Etnk1"]
enh = features["INT_STITCHED_3437"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Pml"]
enh = features["INT_STITCHED_4555"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ooep"]
enh = features["INT_STITCHED_4657"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Rnu3a"]
enh = features["INT_STITCHED_4954"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-2410006H16Rik"]
enh = features["INT_STITCHED_5533"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-E130012A19Rik"]
enh = features["INT_STITCHED_5711"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-LOC108167924"]
enh = features["INT_STITCHED_5711"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Rara"]
enh = features["INT_STITCHED_5719"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Mapt"]
enh = features["INT_STITCHED_5752"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ppp2r5c"]
enh = features["INT_STITCHED_6188_2"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-LOC102638940"]
enh = features["INT_STITCHED_6188_2"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ppp2r5c"]
enh = features["INT_STITCHED_6188"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gm30363"]
enh = features["INT_STITCHED_6709"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Dmtn"]
enh = features["INT_STITCHED_6887"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Tsc22d1"]
enh = features["INT_STITCHED_6904"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Myo10"]
enh = features["INT_STITCHED_7104_2"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Usp7"]
enh = features["INT_STITCHED_7359"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-H2-M5"]
enh = features["INT_STITCHED_7792"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Rab12"]
enh = features["INT_STITCHED_7876_2"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gm36856"]
enh = features["INT_STITCHED_7887"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Gm20544"]
enh = features["INT_STITCHED_8261"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

## SE assigned genes by the original paper:
SEcorr = chr_corr %>%
    filter( Gene %in% query_SE & Enhancer %in% query_SE ) %>%
    filter( emp_p_pos < 0.05 & distance <= 1000000 ) %>%
    filter(pGene >= 0.01 & pEnh >= 0.01 )

Gene	Enhancer	pGene	pEnh	expCotrans	obsCotrans	chisq	emp_p_pos	emp_p_neg	geneLength	enhsLength	geneStrand	enhsStrand	distance
<chr>	<chr>	<dbl>	<dbl>	<dbl>	<dbl>	<dbl>	<dbl>	<dbl>	<int>	<int>	<chr>	<chr>	<int>
GN-Dst	INT_STITCHED_88	0.02229781	0.03136810	0.0006994399	0.007936508	0.074881560	0.000	1.000	10000	4534	+	*	3866
GN-Nr5a2	INT_STITCHED_466	0.05177627	0.05215420	0.0027003495	0.019652305	0.106419118	0.000	1.000	10000	9328	-	*	4524
GN-Etl4	INT_STITCHED_746	0.06727135	0.14436886	0.0097118885	0.019652305	0.010174323	0.000	1.000	10000	25279	+	*	87622
GN-Cbfa2t2	INT_STITCHED_1210	0.08578987	0.05177627	0.0044418792	0.010959940	0.009564670	0.002	0.998	10000	13465	+	*	20294
GN-Zfp704	INT_STITCHED_1392	0.04610733	0.01058201	0.0004879083	0.001889645	0.004027120	0.026	0.974	10000	10001	-	*	29002
GN-B330016D10Rik	INT_STITCHED_2291	0.02040816	0.09561602	0.0019513474	0.004535147	0.003421237	0.049	0.951	6789	29312	+	*	276502
GN-Kazn	INT_STITCHED_2292	0.05442177	0.08541194	0.0046482690	0.012849584	0.014470241	0.000	1.000	10000	16431	-	*	681281
GN-Med13l	INT_STITCHED_2745	0.05668934	0.05631141	0.0031922570	0.007558579	0.005972191	0.018	0.982	10000	17725	+	*	128066
GN-Nanog	INT_STITCHED_3348	0.05933485	0.03325775	0.0019733433	0.010582011	0.037555124	0.000	1.000	10000	17754	+	*	17463
GN-Phc1	INT_STITCHED_3348	0.14550265	0.03325775	0.0048390903	0.009448224	0.004390104	0.043	0.957	10000	17754	-	*	349872
GN-Idh2	INT_STITCHED_3652	0.02872260	0.03250189	0.0009335388	0.003023432	0.004678597	0.018	0.982	10000	28888	-	*	913318
GN-Tead1	INT_STITCHED_3765	0.09108088	0.04497354	0.0040962299	0.009070295	0.006040023	0.014	0.986	10000	2815	+	*	152595
GN-Utf1	INT_STITCHED_3890	0.05668934	0.02456538	0.0013925953	0.019274376	0.229613073	0.000	1.000	10000	3779	+	*	5280
GN-Dppa5a	INT_STITCHED_4657	0.02456538	0.06764928	0.0016618304	0.023809524	0.295168696	0.000	1.000	10000	14043	-	*	8862
GN-Lrrc2	INT_STITCHED_4748	0.02947846	0.01549509	0.0004567713	0.002267574	0.007178660	0.004	0.996	10000	4546	+	*	5327
GN-Derl3	INT_STITCHED_5091	0.01549509	0.02154195	0.0003337944	0.001511716	0.004156747	0.032	0.968	8779	12208	+	*	20906
GN-Rara	INT_STITCHED_5719	0.02721088	0.04383976	0.0011929186	0.004535147	0.009364003	0.002	0.998	10000	6794	+	*	5948
GN-Mapt	INT_STITCHED_5752	0.03061224	0.05442177	0.0016659725	0.011715797	0.060624638	0.000	1.000	10000	14603	+	*	11145
GN-Esrrb	INT_STITCHED_6113	0.01360544	0.12018141	0.0016351212	0.004535147	0.005143443	0.014	0.986	10000	10057	+	*	74887
GN-Ppp2r5c	INT_STITCHED_6188	0.03779289	0.05442177	0.0020567562	0.005668934	0.006343888	0.010	0.990	10000	10256	+	*	11287
GN-Tsc22d1	INT_STITCHED_6904	0.03476946	0.05177627	0.0018002330	0.034013605	0.576426148	0.000	1.000	10000	10591	+	*	5024
GN-Tsc22d1	INT_STITCHED_6906	0.03476946	0.02796674	0.0009723886	0.003023432	0.004326230	0.015	0.985	10000	12300	+	*	115373
GN-Rpap3	INT_STITCHED_7317	0.02343159	0.13454271	0.0031525502	0.006424792	0.003396478	0.047	0.953	10000	19947	-	*	338565
GN-Usp7	INT_STITCHED_7359	0.07936508	0.11602419	0.0092082688	0.016628874	0.005979992	0.041	0.959	10000	11985	-	*	29694
GN-H2-M5	INT_STITCHED_7792	0.05139834	0.07369615	0.0037878593	0.027966742	0.154340046	0.000	1.000	8167	13848	-	*	3732

Gene	Enhancer	pGene	pEnh	expCotrans	obsCotrans	chisq	emp_p_pos	emp_p_neg	geneLength	enhsLength	geneStrand	enhsStrand	distance
<chr>	<chr>	<dbl>	<dbl>	<dbl>	<dbl>	<dbl>	<dbl>	<dbl>	<int>	<int>	<chr>	<chr>	<int>
GN-Dst	INT_STITCHED_88	0.02229781	0.03136810	0.0006994399	0.007936508	0.074881560	0.000	1.000	10000	4534	+	*	3866
GN-Nr5a2	INT_STITCHED_466	0.05177627	0.05215420	0.0027003495	0.019652305	0.106419118	0.000	1.000	10000	9328	-	*	4524
GN-Uck2	INT_STITCHED_556	0.10997732	0.06311413	0.0069411236	0.013983371	0.007144845	0.018	0.982	10000	18183	-	*	1160589
GN-Etl4	INT_STITCHED_746	0.06727135	0.14436886	0.0097118885	0.019652305	0.010174323	0.000	1.000	10000	25279	+	*	87622
GN-Cbfa2t2	INT_STITCHED_1210	0.08578987	0.05177627	0.0044418792	0.010959940	0.009564670	0.002	0.998	10000	13465	+	*	20294
GN-Zfp704	INT_STITCHED_1392	0.04610733	0.01058201	0.0004879083	0.001889645	0.004027120	0.026	0.974	10000	10001	-	*	29002
GN-B330016D10Rik	INT_STITCHED_2291	0.02040816	0.09561602	0.0019513474	0.004535147	0.003421237	0.049	0.951	6789	29312	+	*	276502
GN-Kazn	INT_STITCHED_2292	0.05442177	0.08541194	0.0046482690	0.012849584	0.014470241	0.000	1.000	10000	16431	-	*	681281
GN-Med13l	INT_STITCHED_2745	0.05668934	0.05631141	0.0031922570	0.007558579	0.005972191	0.018	0.982	10000	17725	+	*	128066
GN-Nanog	INT_STITCHED_3348	0.05933485	0.03325775	0.0019733433	0.010582011	0.037555124	0.000	1.000	10000	17754	+	*	17463
GN-Phc1	INT_STITCHED_3348	0.14550265	0.03325775	0.0048390903	0.009448224	0.004390104	0.043	0.957	10000	17754	-	*	349872
GN-Idh2	INT_STITCHED_3652	0.02872260	0.03250189	0.0009335388	0.003023432	0.004678597	0.018	0.982	10000	28888	-	*	913318
GN-Tead1	INT_STITCHED_3765	0.09108088	0.04497354	0.0040962299	0.009070295	0.006040023	0.014	0.986	10000	2815	+	*	152595
GN-Utf1	INT_STITCHED_3890	0.05668934	0.02456538	0.0013925953	0.019274376	0.229613073	0.000	1.000	10000	3779	+	*	5280
GN-Chd9	INT_STITCHED_4179	0.04119426	0.05479970	0.0022574327	0.006046863	0.006361112	0.010	0.990	10000	18391	+	*	1954880
GN-Hmg20a	INT_STITCHED_4555	0.02343159	0.05177627	0.0012132005	0.003779289	0.005427638	0.016	0.984	10000	12081	+	*	1852845
GN-Dppa5a	INT_STITCHED_4657	0.02456538	0.06764928	0.0016618304	0.023809524	0.295168696	0.000	1.000	10000	14043	-	*	8862
GN-Lrrc2	INT_STITCHED_4748	0.02947846	0.01549509	0.0004567713	0.002267574	0.007178660	0.004	0.996	10000	4546	+	*	5327
GN-Derl3	INT_STITCHED_5091	0.01549509	0.02154195	0.0003337944	0.001511716	0.004156747	0.032	0.968	8779	12208	+	*	20906
GN-Rara	INT_STITCHED_5719	0.02721088	0.04383976	0.0011929186	0.004535147	0.009364003	0.002	0.998	10000	6794	+	*	5948
GN-Mapt	INT_STITCHED_5752	0.03061224	0.05442177	0.0016659725	0.011715797	0.060624638	0.000	1.000	10000	14603	+	*	11145
GN-Esrrb	INT_STITCHED_6113	0.01360544	0.12018141	0.0016351212	0.004535147	0.005143443	0.014	0.986	10000	10057	+	*	74887
GN-Ppp2r5c	INT_STITCHED_6188	0.03779289	0.05442177	0.0020567562	0.005668934	0.006343888	0.010	0.990	10000	10256	+	*	11287
GN-Tsc22d1	INT_STITCHED_6904	0.03476946	0.05177627	0.0018002330	0.034013605	0.576426148	0.000	1.000	10000	10591	+	*	5024
GN-Tsc22d1	INT_STITCHED_6906	0.03476946	0.02796674	0.0009723886	0.003023432	0.004326230	0.015	0.985	10000	12300	+	*	115373
GN-Rpap3	INT_STITCHED_7317	0.02343159	0.13454271	0.0031525502	0.006424792	0.003396478	0.047	0.953	10000	19947	-	*	338565
GN-Usp7	INT_STITCHED_7359	0.07936508	0.11602419	0.0092082688	0.016628874	0.005979992	0.041	0.959	10000	11985	-	*	29694
GN-H2-M5	INT_STITCHED_7792	0.05139834	0.07369615	0.0037878593	0.027966742	0.154340046	0.000	1.000	8167	13848	-	*	3732
GN-Kctd16	INT_STITCHED_8140	0.01171580	0.04157218	0.0004870513	0.003023432	0.013208517	0.000	1.000	10000	9781	+	*	1628530

gene = features["GN-Dst"]
enh = features["INT_STITCHED_88"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Nr5a2"]
enh = features["INT_STITCHED_466"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Uck2"]
enh = features["INT_STITCHED_556"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Etl4"]
enh = features["INT_STITCHED_746"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Cbfa2t2"]
enh = features["INT_STITCHED_1210"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Zfp704"]
enh = features["INT_STITCHED_1392"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-B330016D10Rik"]
enh = features["INT_STITCHED_2291"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Kazn"]
enh = features["INT_STITCHED_2292"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Med13l"]
enh = features["INT_STITCHED_2745"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Nanog"]
enh = features["INT_STITCHED_3348"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Phc1"]
enh = features["INT_STITCHED_3348"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Idh2"]
enh = features["INT_STITCHED_3652"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Tead1"]
enh = features["INT_STITCHED_3765"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Utf1"]
enh = features["INT_STITCHED_3890"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Chd9"]
enh = features["INT_STITCHED_4179"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Hmg20a"]
enh = features["INT_STITCHED_4555"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);	

gene = features["GN-Dppa5a"]
enh = features["INT_STITCHED_4657"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Lrrc2"]
enh = features["INT_STITCHED_4748"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Derl3"]
enh = features["INT_STITCHED_5091"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Rara"]
enh = features["INT_STITCHED_5719"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Mapt"]
enh = features["INT_STITCHED_5752"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Esrrb"]
enh = features["INT_STITCHED_6113"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Ppp2r5c"]
enh = features["INT_STITCHED_6188"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Tsc22d1"]
enh = features["INT_STITCHED_6904"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Tsc22d1"]
enh = features["INT_STITCHED_6906"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Rpap3"]
enh = features["INT_STITCHED_7317"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Usp7"]
enh = features["INT_STITCHED_7359"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-H2-M5"]
enh = features["INT_STITCHED_7792"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

gene = features["GN-Kctd16"]
enh = features["INT_STITCHED_8140"];
strand(enh) = "+"
plot_position_pairs( scGRO, gene, enh, dREG );
strand(enh) = "-"
plot_position_pairs( scGRO, gene, enh, dREG );

plot_position_table(scGRO, gene, enh);

marginal_corr = function( reads, gene, enh ) {
    query = features[ c(gene, enh) ];
    qstart= promoters(query, upstream=0, downstream=1);
    
    genepol = subsetByOverlaps(reads, query[1]) %>%
        mutate( gene = distanceToNearest(., qstart[1]) );
    genepol$gstrand = as.character(strand(genepol));
    
    enhpol = subsetByOverlaps(reads, query[2]) %>%
        mutate( enh = distanceToNearest(., qstart[2]) );
    enhpol$estrand = as.character(strand(enhpol));

    genepol = mcols(genepol) %>%
        as.data.frame %>%
        filter(cellID %in% enhpol$cellID) %>%
        select(cellID, gene.distance, gstrand);
    
    enhpol  = mcols(enhpol) %>%
        as.data.frame %>%
        filter(cellID %in% genepol$cellID) %>%
        select(cellID, enh.distance, estrand);

    p = left_join(genepol, enhpol, by="cellID") %>%
            ggplot(aes(x=gene.distance/1000, y=enh.distance/1000)) +
            geom_point() +
            xlim(0, 50) +
            ylim(0, 20) +
            xlab("Gene distance transcribed (kb)") +
            ylab("Enhancer distance transcribed (kb)") +
            theme(legend.position="none");

    ggMarginal(p, type="histogram")
}

marginal_corr( scGRO, "GN-Sox2", "Sox2_105kbDn_pl" );

marginal_corr( scGRO, "GN-Pou5f1", "Pou5f1_25kbUp_mn" );
